refactor codebase for clarity, improve inheritance and composition hierarchy, introduce strict mode for client-side schema checking
