+++ integration/test_collection.py
@@ -16,9 +16,10 @@ from weaviate.collection.classes import (
     PQEncoderConfigUpdate,
     PQEncoderType,
     PQEncoderDistribution,
-    ReferenceDataType,
     ReferenceProperty,
+    ReferencePropertyMultiTarget,
     ReferenceTo,
+    ReferenceToMultiTarget,
     StopwordsUpdate,
     MultiTenancyConfig,
     StopwordsPreset,
@@ -26,7 +27,7 @@ from weaviate.collection.classes import (
     VectorIndexConfigUpdate,
     Vectorizer,
 )
-from weaviate.collection.grpc import HybridFusion, LinkTo, MetadataQuery
+from weaviate.collection.grpc import HybridFusion, LinkTo, LinkToMultiTarget, MetadataQuery
 
 
 @pytest.fixture(scope="module")
@@ -248,36 +249,34 @@ def test_reference_add_delete_replace(client: weaviate.Client):
     uuid_to = ref_collection.data.insert(data={})
     collection_config = CollectionConfig(
         name="SomethingElse",
-        properties=[
-            ReferenceProperty(name="ref", reference=ReferenceDataType(collections="RefClass2"))
-        ],
+        properties=[ReferenceProperty(name="ref", target_collection="RefClass2")],
         vectorizer=Vectorizer.NONE,
     )
     collection = client.collection.create(collection_config)
 
     uuid_from1 = collection.data.insert({}, uuid.uuid4())
-    uuid_from2 = collection.data.insert({"ref": ReferenceTo(reference_uuids=uuid_to)}, uuid.uuid4())
+    uuid_from2 = collection.data.insert({"ref": ReferenceTo(uuids=uuid_to)}, uuid.uuid4())
     collection.data.reference_add(
-        from_uuid=uuid_from1, from_property="ref", ref=ReferenceTo(reference_uuids=uuid_to)
+        from_uuid=uuid_from1, from_property="ref", ref=ReferenceTo(uuids=uuid_to)
     )
     objects = collection.data.get()
     for obj in objects:
         assert str(uuid_to) in "".join([ref["beacon"] for ref in obj.data["ref"]])
 
     collection.data.reference_delete(
-        from_uuid=uuid_from1, from_property="ref", ref=ReferenceTo(reference_uuids=uuid_to)
+        from_uuid=uuid_from1, from_property="ref", ref=ReferenceTo(uuids=uuid_to)
     )
     assert len(collection.data.get_by_id(uuid_from1).data["ref"]) == 0
 
     collection.data.reference_add(
-        from_uuid=uuid_from2, from_property="ref", ref=ReferenceTo(reference_uuids=uuid_to)
+        from_uuid=uuid_from2, from_property="ref", ref=ReferenceTo(uuids=uuid_to)
     )
     obj = collection.data.get_by_id(uuid_from2)
     assert len(obj.data["ref"]) == 2
     assert str(uuid_to) in "".join([ref["beacon"] for ref in obj.data["ref"]])
 
     collection.data.reference_replace(
-        from_uuid=uuid_from2, from_property="ref", ref=ReferenceTo(reference_uuids=[])
+        from_uuid=uuid_from2, from_property="ref", ref=ReferenceTo(uuids=[])
     )
     assert len(collection.data.get_by_id(uuid_from2).data["ref"]) == 0
 
@@ -475,27 +474,25 @@ def test_mono_references_grcp(client: weaviate.Client):
             name="B",
             properties=[
                 Property(name="Name", data_type=DataType.TEXT),
-                ReferenceProperty(name="ref", reference=ReferenceDataType(collections="A")),
+                ReferenceProperty(name="ref", target_collection="A"),
             ],
             vectorizer=Vectorizer.NONE,
         )
     )
-    uuid_B = B.data.insert({"Name": "B", "ref": ReferenceTo(reference_uuids=uuid_A1)})
-    B.data.reference_add(
-        from_uuid=uuid_B, from_property="ref", ref=ReferenceTo(reference_uuids=uuid_A2)
-    )
+    uuid_B = B.data.insert({"Name": "B", "ref": ReferenceTo(uuids=uuid_A1)})
+    B.data.reference_add(from_uuid=uuid_B, from_property="ref", ref=ReferenceTo(uuids=uuid_A2))
 
     C = client.collection.create(
         CollectionConfig(
             name="C",
             properties=[
                 Property(name="Name", data_type=DataType.TEXT),
-                ReferenceProperty(name="ref", reference=ReferenceDataType(collections="B")),
+                ReferenceProperty(name="ref", target_collection="B"),
             ],
             vectorizer=Vectorizer.NONE,
         )
     )
-    C.data.insert({"Name": "find me", "ref": ReferenceTo(reference_uuids=uuid_B)})
+    C.data.insert({"Name": "find me", "ref": ReferenceTo(uuids=uuid_B)})
 
     objects = C.query.bm25_flat(
         query="find",
@@ -553,25 +550,25 @@ def test_multi_references_grcp(client: weaviate.Client):
             name="C",
             properties=[
                 Property(name="Name", data_type=DataType.TEXT),
-                ReferenceProperty(name="ref", reference=ReferenceDataType(collections=["A", "B"])),
+                ReferencePropertyMultiTarget(name="ref", target_collections=["A", "B"]),
             ],
             vectorizer=Vectorizer.NONE,
         )
     )
     C.data.insert(
-        {"Name": "first", "ref": ReferenceTo(reference_uuids=uuid_A, which_collection="A")}
+        {"Name": "first", "ref": ReferenceToMultiTarget(uuids=uuid_A, target_collection="A")}
     )
     C.data.insert(
-        {"Name": "second", "ref": ReferenceTo(reference_uuids=uuid_B, which_collection="B")}
+        {"Name": "second", "ref": ReferenceToMultiTarget(uuids=uuid_B, target_collection="B")}
     )
 
     objects = C.query.bm25_flat(
         query="first",
         return_properties=[
             "name",
-            LinkTo(
+            LinkToMultiTarget(
                 link_on="ref",
-                which_collection="A",
+                target_collection="A",
                 properties=["name"],
                 metadata=MetadataQuery(uuid=True, last_update_time_unix=True),
             ),
@@ -585,9 +582,9 @@ def test_multi_references_grcp(client: weaviate.Client):
         query="second",
         return_properties=[
             "name",
-            LinkTo(
+            LinkToMultiTarget(
                 link_on="ref",
-                which_collection="B",
+                target_collection="B",
                 properties=[
                     "name",
                 ],
