+++ test/collection/test_classes.py
@@ -1,34 +1,34 @@
 import pytest
 import uuid
-from weaviate.collection.classes import ReferenceTo, _ReferenceDataType
+from weaviate.collection.classes import ReferenceTo, ReferenceToMultiTarget, _ReferenceDataType
 
 
-def test_raises_error_when_which_collection_required_and_absent():
+def test_raises_error_when_reference_to_used_incorrectly():
     ref_dtype = _ReferenceDataType(
         collections=["Foo", "Bar"],
     )
     ref_to = ReferenceTo(
-        reference_uuids=uuid.uuid4(),
+        uuids=uuid.uuid4(),
     )
     with pytest.raises(ValueError) as error:
-        ref_to.to_beacons(ref_dtype)
+        ref_to.to_beacons_strict(ref_dtype)
         assert (
             error
-            == "which_collection must be specified when using a reference property with multiple target collections"
+            == "Can only use ReferenceTo on a reference property with a single target collection, use ReferenceToMultiTarget instead"
         )
 
 
-def test_raises_error_when_which_collection_required_and_wrong():
+def test_raises_error_when_target_collection_required_and_wrong():
     ref_dtype = _ReferenceDataType(
         collections=["Foo", "Bar"],
     )
-    ref_to = ReferenceTo(
-        reference_uuids=uuid.uuid4(),
-        which_collection="Baz",
+    ref_to = ReferenceToMultiTarget(
+        uuids=uuid.uuid4(),
+        target_collection="Baz",
     )
     with pytest.raises(ValueError) as error:
-        ref_to.to_beacons(ref_dtype)
+        ref_to.to_beacons_strict(ref_dtype)
         assert (
             error
-            == 'which_collection must be one of ["Foo", "Bar"] since these are the target collections specified in the reference property of this collection but got "Baz" instead'
+            == 'target_collection must be one of ["Foo", "Bar"] since these are the collections specified in the reference property of this collection but got "Baz" instead'
         )
