+++ weaviate/collection/tenants.py
@@ -1,0 +1,99 @@
+from typing import Dict, Any, List
+
+from requests.exceptions import ConnectionError as RequestsConnectionError
+
+from weaviate.collection.classes import Tenant
+from weaviate.connect import Connection
+from weaviate.exceptions import UnexpectedStatusCodeException
+
+
+class _Tenants:
+    """
+    Represents all the CRUD methods available on a collection's multi-tenancy specification within Weaviate. The
+    collection must have been created with multi-tenancy enabled in order to use any of these methods. This class
+    should not be instantiated directly, but is available as a property of the `Collection` class under
+    the `collection.tenants` class attribute.
+    """
+
+    def __init__(self, connection: Connection, name: str) -> None:
+        self._connection = connection
+        self.name = name
+
+    def add(self, tenants: List[Tenant]) -> None:
+        """Add the specified tenants to a collection in Weaviate.
+
+        The collection must have been created with multi-tenancy enabled.
+
+        Parameters:
+        - `tenants`: List of Tenants.
+
+        Raises:
+        - `requests.ConnectionError`
+            - If the network connection to Weaviate fails.
+        - `weaviate.UnexpectedStatusCodeException`
+            - If Weaviate reports a non-OK status.
+        """
+
+        loaded_tenants = [{"name": tenant.name} for tenant in tenants]
+
+        path = "/schema/" + self.name + "/tenants"
+        try:
+            response = self._connection.post(path=path, weaviate_object=loaded_tenants)
+        except RequestsConnectionError as conn_err:
+            raise RequestsConnectionError(
+                f"Collection tenants may not have been added properly for {self.name}"
+            ) from conn_err
+        if response.status_code != 200:
+            raise UnexpectedStatusCodeException(f"Add collection tenants for {self.name}", response)
+
+    def remove(self, tenants: List[str]) -> None:
+        """Remove the specified tenants from a collection in Weaviate.
+
+        The collection must have been created with multi-tenancy enabled.
+
+        Parameters:
+        - `tenants`: List of tenant names to remove from the given class.
+
+        Raises:
+        - `TypeError`
+            - If 'tenants' has not the correct type.
+        - `requests.ConnectionError`
+            - If the network connection to Weaviate fails.
+        - `weaviate.UnexpectedStatusCodeException`
+            - If Weaviate reports a non-OK status.
+        """
+        path = "/schema/" + self.name + "/tenants"
+        try:
+            response = self._connection.delete(path=path, weaviate_object=tenants)
+        except RequestsConnectionError as conn_err:
+            raise RequestsConnectionError(
+                f"Collection tenants may not have been deleted for {self.name}"
+            ) from conn_err
+        if response.status_code != 200:
+            raise UnexpectedStatusCodeException(
+                f"Delete collection tenants for {self.name}", response
+            )
+
+    def get(self) -> List[Tenant]:
+        """Return all tenants currently associated with a collection in Weaviate.
+
+        The collection must have been created with multi-tenancy enabled.
+
+        Raises:
+        - `requests.ConnectionError`
+            - If the network connection to Weaviate fails.
+        - `weaviate.UnexpectedStatusCodeException`
+            - If Weaviate reports a non-OK status.
+        """
+        path = "/schema/" + self.name + "/tenants"
+        try:
+            response = self._connection.get(path=path)
+        except RequestsConnectionError as conn_err:
+            raise RequestsConnectionError(
+                f"Could not get collection tenants for {self.name}"
+            ) from conn_err
+        if response.status_code != 200:
+            raise UnexpectedStatusCodeException(f"Get collection tenants for {self.name}", response)
+
+        tenant_resp: List[Dict[str, Any]] = response.json()
+        return [Tenant(**tenant) for tenant in tenant_resp]
