+++ weaviate/weaviate_classes.py
@@ -9,8 +9,11 @@ from weaviate.collection.classes import (
     InvertedIndexConfigCreate,
     InvertedIndexConfigUpdate,
     Property,
+    ReferenceDataType,
     ReferenceProperty,
+    ReferencePropertyMultiTarget,
     ReferenceTo,
+    ReferenceToMultiTarget,
     ShardingConfigCreate,
     StopwordsCreate,
     StopwordsUpdate,
@@ -27,6 +30,7 @@ from weaviate.collection.grpc import (
     HybridFusion,
     HybridOptions,
     LinkTo,
+    LinkToMultiTarget,
     MetadataQuery,
     NearObjectOptions,
     NearVectorOptions,
@@ -48,12 +52,16 @@ __all__ = [
     "InvertedIndexConfigCreate",
     "InvertedIndexConfigUpdate",
     "LinkTo",
+    "LinkToMultiTarget",
     "MetadataQuery",
     "NearObjectOptions",
     "NearVectorOptions",
+    "ReferenceDataType",
     "ReferenceProperty",
+    "ReferencePropertyMultiTarget",
     "Property",
     "ReferenceTo",
+    "ReferenceToMultiTarget",
     "ReturnValues",
     "ShardingConfigCreate",
     "StopwordsCreate",
